Converting windows security group test to stop using v2
* in the process moving/creating some convenience helpers

Authored-by: Renee Chu <rchu@pivotal.io>
