+++ assets/catnip/manifest.yml
@@ -1,0 +1,6 @@
+---
+applications:
+- name: catnip
+  processes:
+  - type: worker
+    command: start-worker.sh
