+++ windows/running_security_groups_win.go
@@ -3,6 +3,7 @@ package windows
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/cloudfoundry/cf-acceptance-tests/helpers/v3_helpers"
 	"io/ioutil"
 	"os"
 
@@ -20,13 +21,7 @@ import (
 	"github.com/cloudfoundry/cf-acceptance-tests/helpers/random_name"
 )
 
-type AppsResponse struct {
-	Resources []struct {
-		Metadata struct {
-			Url string
-		}
-	}
-}
+
 
 type StatsResponse map[string]struct {
 	Stats struct {
@@ -60,13 +55,12 @@ func warmUpRequest(appName string) {
 }
 
 func getAppHostIpAndPort(appName string) (string, int) {
-	var appsResponse AppsResponse
-	cfResponse := cf.Cf("curl", fmt.Sprintf("/v2/apps?q=name:%s", appName)).Wait().Out.Contents()
-	json.Unmarshal(cfResponse, &appsResponse)
-	serverAppUrl := appsResponse.Resources[0].Metadata.Url
+	app := v3_helpers.GetApp(appName)
+	processGuid := v3_helpers.GetProcessGuidForType(app.GUID, "web")[0]
+	processUrl := fmt.Sprintf("v3/processes/%s", processGuid)
 
 	var statsResponse StatsResponse
-	cfResponse = cf.Cf("curl", fmt.Sprintf("%s/stats", serverAppUrl)).Wait().Out.Contents()
+	cfResponse := cf.Cf("curl", fmt.Sprintf("%s/stats", processUrl)).Wait().Out.Contents()
 	json.Unmarshal(cfResponse, &statsResponse)
 
 	return statsResponse["0"].Stats.Host, statsResponse["0"].Stats.Port
@@ -147,7 +141,7 @@ func assertNetworkingPreconditions(clientAppName string, privateHost string, pri
 	Expect(noraCurlResponse.ReturnCode).NotTo(Equal(0), "Expected default running security groups not to allow internal communication between app containers. Configure your running security groups to not allow traffic on internal networks, or disable this test by setting 'include_security_groups' to 'false' in '"+os.Getenv("CONFIG")+"'.")
 }
 
-var _ = WindowsDescribe("WINDOWS: App Instance Networking", func() {
+var _ = FDescribe("WINDOWS: App Instance Networking", func() {
 	SecurityGroupsDescribe("WINDOWS: Using container-networking and running security-groups", func() {
 		var serverAppName, clientAppName, privateHost, orgName, spaceName, securityGroupName string
 		var privatePort int
